broke down post template, with edit/comment forms added
