+++ js/modules/functionality/modal.mjs
@@ -9,15 +9,9 @@ const closeModalBtn = document.querySelector(".close-modal-btn");
  * @param {Object} data a single post object 
  */
 export function openPostModal(data){
-  /*round about way of doing it but I couldn't find a reason for bug
-    in createAPost function, where boolean passed changed after first
-    post feed item was created*/
-  if(body.classList.contains("modal-open")){
-  } else {
-    body.classList.add("modal-open");
-    renderPost(data, postContent, true);
-    closeModalBtn.addEventListener("click", closePostModal);
-  }
+  body.classList.add("modal-open");
+  renderPost(data, postContent, true);
+  closeModalBtn.addEventListener("click", closePostModal);
 }
 
 /**
